refactor: move H into Heading
