+++ src/components/Article/Article.mdx
@@ -1,19 +1,19 @@
 ```typescript jsx
 <Box>
-  <H>This will be a h2</H>
+  <Heading>This will be a h2</Heading>
   <Article>
-    <H>This will be a h3</H>
+    <Heading>This will be a h3</Heading>
     <Article>
-      <H>This will be a h4</H>
+      <Heading>This will be a h4</Heading>
 
       <Article>
-        <H>This will be a h5</H>
+        <Heading>This will be a h5</Heading>
 
         <Article>
-          <H>This will be a h6</H>
+          <Heading>This will be a h6</Heading>
 
           <Article>
-            <H>This will still be a h6</H>
+            <Heading>This will still be a h6</Heading>
           </Article>
         </Article>
       </Article>
