+++ src/components/Article/Article.tsx
@@ -3,7 +3,7 @@ import { BoxProps } from '../Box';
 import Section from '../Section';
 
 /**
- * Increases the heading level of all `H` components inside it by 1.
+ * Increases the heading level of all `Heading` components inside it by 1.
  *
  * **NOTE**: this component is functionally the same as `<Section />`, but semantically
  * it renders an `<article>` tag instead of a `<section>` tag.
