--- src/components/H/H.mdx
@@ -1,23 +1,0 @@
-```typescript jsx
-<Box>
-  <H>This will be a h2</H>
-  <Section>
-    <H>This will be a h3</H>
-    <Section>
-      <H>This will be a h4</H>
-
-      <Section>
-        <H>This will be a h5</H>
-
-        <Section>
-          <H>This will be a h6</H>
-
-          <Section>
-            <H>This will still be a h6</H>
-          </Section>
-        </Section>
-      </Section>
-    </Section>
-  </Section>
-</Box>
-```
