--- src/components/H/index.ts
@@ -1,1 +1,0 @@
-export { default, HeadingLevelContext } from './H';
