+++ src/components/Heading/index.tsx
@@ -1,2 +1,2 @@
-export { default } from './Heading';
+export { default, HeadingLevelContext } from './Heading';
 export type { HeadingProps } from './Heading';
