+++ src/components/Section/Section.mdx
@@ -1,19 +1,19 @@
 ```typescript jsx
 <Box>
-  <H>This will be a h2</H>
+  <Heading>This will be a h2</Heading>
   <Section>
-    <H>This will be a h3</H>
+    <Heading>This will be a h3</Heading>
     <Section>
-      <H>This will be a h4</H>
+      <Heading>This will be a h4</Heading>
 
       <Section>
-        <H>This will be a h5</H>
+        <Heading>This will be a h5</Heading>
 
         <Section>
-          <H>This will be a h6</H>
+          <Heading>This will be a h6</Heading>
 
           <Section>
-            <H>This will still be a h6</H>
+            <Heading>This will still be a h6</Heading>
           </Section>
         </Section>
       </Section>
